Glowing Mushroom Entity
The entity emits a light source.
Player shoots it to inflict damage, once it gets to 0, the mushroom entity is destroyed, spawning a healing orb that heals player.
PlayerHealth script is updated to include healing function.
Glowing Mushroom has a health script
Healing Orb has a script that handles collision with player.