Fly Queen & Status Mechanics Initial Draft
Finished Fly Queen and Fly Larvae behavior scripts. PlayerHealth script has been modified to account for a vulnerability effect and defense on the Player. There is a simple StatusMechanics script as well in case the officers decide to build off it.

NOTE: If the larvae spawnpoints are in the Queen, then set the larvae collider isTrigger to true; otherwise, set it to false. This is due to the Fly Queen possessing a collider with isTrigger to false.