Add LevelChanger gameobject and script to fade out and into scenes
The LevelChanger object fades into any scene it is in unless another script within the scene prevents the fade in. Upon transitioning to another scene, the LevelChanger script is called to fade out and then transition to the desired scene after the fade out animation is complete. The fade out animation does not function for CopilotUI scene at the moment.